# Update k8s versions from x.26.2 to x.27.3