--- a/kubernetes/kubernetes.checkout
+++ b/kubernetes/kubernetes.checkout
@@ -2,4 +2,5 @@ staging/src/k8s.io/api/core/v1/generated.proto
 staging/src/k8s.io/apimachinery/pkg/api/resource/generated.proto
 staging/src/k8s.io/apimachinery/pkg/apis/meta/v1/generated.proto
 staging/src/k8s.io/apimachinery/pkg/runtime/generated.proto
+staging/src/k8s.io/apimachinery/pkg/runtime/schema/generated.proto
 staging/src/k8s.io/apimachinery/pkg/util/intstr/generated.proto