--- a/kubernetes/proto/tools/go.mod
+++ b/kubernetes/proto/tools/go.mod
@@ -3,7 +3,7 @@ module sigs.k8s.io/kube-scheduler-wasm-extension/kubernetes/proto/tools
 go 1.20
 
 require (
-	github.com/knqyf263/go-plugin v0.7.0
+	github.com/knqyf263/go-plugin v0.8.0
 	google.golang.org/protobuf v1.30.0
 )