--- a/scheduler/go.mod
+++ b/scheduler/go.mod
@@ -4,56 +4,57 @@ go 1.20
 
 replace (
 	google.golang.org/grpc => google.golang.org/grpc v1.38.0
-	k8s.io/api => k8s.io/api v0.26.2
-	k8s.io/apiextensions-apiserver => k8s.io/apiextensions-apiserver v0.26.2
-	k8s.io/apimachinery => k8s.io/apimachinery v0.26.2
-	k8s.io/apiserver => k8s.io/apiserver v0.26.2
-	k8s.io/cli-runtime => k8s.io/cli-runtime v0.26.2
-	k8s.io/client-go => k8s.io/client-go v0.26.2
-	k8s.io/cloud-provider => k8s.io/cloud-provider v0.26.2
-	k8s.io/cluster-bootstrap => k8s.io/cluster-bootstrap v0.26.2
-	k8s.io/code-generator => k8s.io/code-generator v0.26.2
-	k8s.io/component-base => k8s.io/component-base v0.26.2
-	k8s.io/component-helpers => k8s.io/component-helpers v0.26.2
-	k8s.io/controller-manager => k8s.io/controller-manager v0.26.2
-	k8s.io/cri-api => k8s.io/cri-api v0.26.2
-	k8s.io/csi-translation-lib => k8s.io/csi-translation-lib v0.26.2
-	k8s.io/dynamic-resource-allocation => k8s.io/dynamic-resource-allocation v0.26.2
-	k8s.io/kms => k8s.io/kms v0.26.1
-	k8s.io/kube-aggregator => k8s.io/kube-aggregator v0.26.2
-	k8s.io/kube-controller-manager => k8s.io/kube-controller-manager v0.26.2
-	k8s.io/kube-proxy => k8s.io/kube-proxy v0.26.2
-	k8s.io/kube-scheduler => k8s.io/kube-scheduler v0.26.2
-	k8s.io/kubectl => k8s.io/kubectl v0.26.2
-	k8s.io/kubelet => k8s.io/kubelet v0.26.2
-	k8s.io/legacy-cloud-providers => k8s.io/legacy-cloud-providers v0.26.2
-	k8s.io/metrics => k8s.io/metrics v0.26.2
-	k8s.io/mount-utils => k8s.io/mount-utils v0.26.2
-	k8s.io/pod-security-admission => k8s.io/pod-security-admission v0.26.2
-	k8s.io/sample-apiserver => k8s.io/sample-apiserver v0.26.2
+	k8s.io/api => k8s.io/api v0.27.3
+	k8s.io/apiextensions-apiserver => k8s.io/apiextensions-apiserver v0.27.3
+	k8s.io/apimachinery => k8s.io/apimachinery v0.27.3
+	k8s.io/apiserver => k8s.io/apiserver v0.27.3
+	k8s.io/cli-runtime => k8s.io/cli-runtime v0.27.3
+	k8s.io/client-go => k8s.io/client-go v0.27.3
+	k8s.io/cloud-provider => k8s.io/cloud-provider v0.27.3
+	k8s.io/cluster-bootstrap => k8s.io/cluster-bootstrap v0.27.3
+	k8s.io/code-generator => k8s.io/code-generator v0.27.3
+	k8s.io/component-base => k8s.io/component-base v0.27.3
+	k8s.io/component-helpers => k8s.io/component-helpers v0.27.3
+	k8s.io/controller-manager => k8s.io/controller-manager v0.27.3
+	k8s.io/cri-api => k8s.io/cri-api v0.27.3
+	k8s.io/csi-translation-lib => k8s.io/csi-translation-lib v0.27.3
+	k8s.io/dynamic-resource-allocation => k8s.io/dynamic-resource-allocation v0.27.3
+	k8s.io/kms => k8s.io/kms v0.27.3
+	k8s.io/kube-aggregator => k8s.io/kube-aggregator v0.27.3
+	k8s.io/kube-controller-manager => k8s.io/kube-controller-manager v0.27.3
+	k8s.io/kube-proxy => k8s.io/kube-proxy v0.27.3
+	k8s.io/kube-scheduler => k8s.io/kube-scheduler v0.27.3
+	k8s.io/kubectl => k8s.io/kubectl v0.27.3
+	k8s.io/kubelet => k8s.io/kubelet v0.27.3
+	k8s.io/legacy-cloud-providers => k8s.io/legacy-cloud-providers v0.27.3
+	k8s.io/metrics => k8s.io/metrics v0.27.3
+	k8s.io/mount-utils => k8s.io/mount-utils v0.27.3
+	k8s.io/pod-security-admission => k8s.io/pod-security-admission v0.27.3
+	k8s.io/sample-apiserver => k8s.io/sample-apiserver v0.27.3
 	sigs.k8s.io/kube-scheduler-wasm-extension/kubernetes/proto => ../kubernetes/proto
 )
 
 require (
-	github.com/google/uuid v1.1.2
+	github.com/google/uuid v1.3.0
 	github.com/tetratelabs/wazero v1.2.1
-	k8s.io/api v0.26.2
-	k8s.io/apimachinery v0.26.2
-	k8s.io/component-base v0.26.2
-	k8s.io/kubectl v0.26.2
-	k8s.io/kubernetes v1.26.2
+	k8s.io/api v0.27.3
+	k8s.io/apimachinery v0.27.3
+	k8s.io/component-base v0.27.3
+	k8s.io/kubectl v0.27.3
+	k8s.io/kubernetes v1.27.3
 )
 
 require (
 	github.com/Azure/go-ansiterm v0.0.0-20210617225240-d185dfc1b5a1 // indirect
 	github.com/NYTimes/gziphandler v1.1.1 // indirect
 	github.com/antlr/antlr4/runtime/Go/antlr v1.4.10 // indirect
+	github.com/asaskevich/govalidator v0.0.0-20190424111038-f61b66f89f4a // indirect
 	github.com/beorn7/perks v1.0.1 // indirect
 	github.com/blang/semver/v4 v4.0.0 // indirect
 	github.com/cenkalti/backoff/v4 v4.1.3 // indirect
 	github.com/cespare/xxhash/v2 v2.1.2 // indirect
 	github.com/coreos/go-semver v0.3.0 // indirect
-	github.com/coreos/go-systemd/v22 v22.3.2 // indirect
+	github.com/coreos/go-systemd/v22 v22.4.0 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/docker/distribution v2.8.1+incompatible // indirect
 	github.com/emicklei/go-restful/v3 v3.9.0 // indirect
@@ -62,12 +63,12 @@ require (
 	github.com/fsnotify/fsnotify v1.6.0 // indirect
 	github.com/go-logr/logr v1.2.3 // indirect
 	github.com/go-logr/stdr v1.2.2 // indirect
-	github.com/go-openapi/jsonpointer v0.19.5 // indirect
-	github.com/go-openapi/jsonreference v0.20.0 // indirect
-	github.com/go-openapi/swag v0.19.14 // indirect
+	github.com/go-openapi/jsonpointer v0.19.6 // indirect
+	github.com/go-openapi/jsonreference v0.20.1 // indirect
+	github.com/go-openapi/swag v0.22.3 // indirect
 	github.com/gogo/protobuf v1.3.2 // indirect
 	github.com/golang/groupcache v0.0.0-20210331224755-41bb18bfe9da // indirect
-	github.com/golang/protobuf v1.5.2 // indirect
+	github.com/golang/protobuf v1.5.3 // indirect
 	github.com/google/cel-go v0.12.6 // indirect
 	github.com/google/gnostic v0.5.7-v3refs // indirect
 	github.com/google/go-cmp v0.5.9 // indirect
@@ -78,10 +79,11 @@ require (
 	github.com/inconshreveable/mousetrap v1.0.1 // indirect
 	github.com/josharian/intern v1.0.0 // indirect
 	github.com/json-iterator/go v1.1.12 // indirect
-	github.com/mailru/easyjson v0.7.6 // indirect
+	github.com/mailru/easyjson v0.7.7 // indirect
 	github.com/matttproud/golang_protobuf_extensions v1.0.2 // indirect
+	github.com/mitchellh/mapstructure v1.4.1 // indirect
 	github.com/moby/sys/mountinfo v0.6.2 // indirect
-	github.com/moby/term v0.0.0-20220808134915-39b0c02b01ae // indirect
+	github.com/moby/term v0.0.0-20221205130635-1aeaba878587 // indirect
 	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
 	github.com/modern-go/reflect2 v1.0.2 // indirect
 	github.com/munnerz/goautoneg v0.0.0-20191010083416-a7dc8b61c822 // indirect
@@ -95,11 +97,11 @@ require (
 	github.com/spf13/cobra v1.6.0 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
 	github.com/stoewer/go-strcase v1.2.0 // indirect
-	go.etcd.io/etcd/api/v3 v3.5.5 // indirect
-	go.etcd.io/etcd/client/pkg/v3 v3.5.5 // indirect
-	go.etcd.io/etcd/client/v3 v3.5.5 // indirect
+	go.etcd.io/etcd/api/v3 v3.5.7 // indirect
+	go.etcd.io/etcd/client/pkg/v3 v3.5.7 // indirect
+	go.etcd.io/etcd/client/v3 v3.5.7 // indirect
 	go.opentelemetry.io/contrib/instrumentation/google.golang.org/grpc/otelgrpc v0.35.0 // indirect
-	go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp v0.35.0 // indirect
+	go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp v0.35.1 // indirect
 	go.opentelemetry.io/otel v1.10.0 // indirect
 	go.opentelemetry.io/otel/exporters/otlp/internal/retry v1.10.0 // indirect
 	go.opentelemetry.io/otel/exporters/otlp/otlptrace v1.10.0 // indirect
@@ -112,35 +114,37 @@ require (
 	go.uber.org/multierr v1.6.0 // indirect
 	go.uber.org/zap v1.19.0 // indirect
 	golang.org/x/crypto v0.1.0 // indirect
-	golang.org/x/net v0.7.0 // indirect
+	golang.org/x/net v0.8.0 // indirect
 	golang.org/x/oauth2 v0.0.0-20220223155221-ee480838109b // indirect
-	golang.org/x/sync v0.0.0-20220722155255-886fb9371eb4 // indirect
-	golang.org/x/sys v0.5.0 // indirect
-	golang.org/x/term v0.5.0 // indirect
-	golang.org/x/text v0.7.0 // indirect
+	golang.org/x/sync v0.1.0 // indirect
+	golang.org/x/sys v0.6.0 // indirect
+	golang.org/x/term v0.6.0 // indirect
+	golang.org/x/text v0.8.0 // indirect
 	golang.org/x/time v0.0.0-20220210224613-90d013bbcef8 // indirect
 	google.golang.org/appengine v1.6.7 // indirect
 	google.golang.org/genproto v0.0.0-20220502173005-c8bf987b8c21 // indirect
-	google.golang.org/grpc v1.49.0 // indirect
+	google.golang.org/grpc v1.51.0 // indirect
 	google.golang.org/protobuf v1.28.1 // indirect
 	gopkg.in/inf.v0 v0.9.1 // indirect
 	gopkg.in/natefinch/lumberjack.v2 v2.0.0 // indirect
 	gopkg.in/yaml.v2 v2.4.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
-	k8s.io/apiserver v0.26.2 // indirect
-	k8s.io/client-go v0.26.2 // indirect
+	k8s.io/apiserver v0.27.3 // indirect
+	k8s.io/client-go v0.27.3 // indirect
 	k8s.io/cloud-provider v0.0.0 // indirect
-	k8s.io/component-helpers v0.26.2 // indirect
+	k8s.io/component-helpers v0.27.3 // indirect
+	k8s.io/controller-manager v0.27.3 // indirect
 	k8s.io/csi-translation-lib v0.0.0 // indirect
 	k8s.io/dynamic-resource-allocation v0.0.0 // indirect
-	k8s.io/klog/v2 v2.80.1 // indirect
-	k8s.io/kms v0.26.2 // indirect
-	k8s.io/kube-openapi v0.0.0-20221012153701-172d655c2280 // indirect
+	k8s.io/klog/v2 v2.90.1 // indirect
+	k8s.io/kms v0.27.3 // indirect
+	k8s.io/kube-openapi v0.0.0-20230501164219-8b0f38b5fd1f // indirect
 	k8s.io/kube-scheduler v0.0.0 // indirect
+	k8s.io/kubelet v0.27.3 // indirect
 	k8s.io/mount-utils v0.0.0 // indirect
-	k8s.io/utils v0.0.0-20221107191617-1a15be271d1d // indirect
-	sigs.k8s.io/apiserver-network-proxy/konnectivity-client v0.0.35 // indirect
-	sigs.k8s.io/json v0.0.0-20220713155537-f223a00ba0e2 // indirect
+	k8s.io/utils v0.0.0-20230209194617-a36077c30491 // indirect
+	sigs.k8s.io/apiserver-network-proxy/konnectivity-client v0.1.2 // indirect
+	sigs.k8s.io/json v0.0.0-20221116044647-bc3834ca7abd // indirect
 	sigs.k8s.io/structured-merge-diff/v4 v4.2.3 // indirect
 	sigs.k8s.io/yaml v1.3.0 // indirect
 )