#### What type of PR is this?

/kind cleanup

#### What this PR does / why we need it:

Before, we were using inconsistent versions of k8s (mostly x.26.2, but some x.26.1). Also, we used an arbitrary commit from master to generate proto sources. This updates all k8s dependencies to x.27.3 and uses the 1.27.3 commit to generate protos.

This also uses a newer version of the source generation plugin.

#### Which issue(s) this PR fixes:

NONE

#### Special notes for your reviewer:

#### Does this PR introduce a user-facing change?

NONE

#### What are the benchmark results of this change?

```benchstat
goos: darwin
goarch: arm64
pkg: sigs.k8s.io/kube-scheduler-wasm-extension/internal/e2e
                                                        │ before.txt  │             after.txt             │
                                                        │   sec/op    │   sec/op     vs base              │
PluginPreFilter/noop-wat/params:_small-12                 133.7n ± 1%   133.2n ± 0%  -0.41% (p=0.006 n=6)
PluginPreFilter/noop-wat/params:_real-12                  134.3n ± 1%   133.2n ± 1%       ~ (p=0.056 n=6)
PluginPreFilter/noop/params:_small-12                     157.6n ± 1%   156.3n ± 0%  -0.82% (p=0.002 n=6)
PluginPreFilter/noop/params:_real-12                      157.2n ± 1%   157.0n ± 0%       ~ (p=0.106 n=6)
PluginPreFilter/test/params:_small-12                     3.835µ ± 0%   3.954µ ± 0%  +3.10% (p=0.002 n=6)
PluginPreFilter/test/params:_real-12                      45.02µ ± 1%   46.03µ ± 1%  +2.23% (p=0.002 n=6)
PluginFilter/noop-wat/params:_small-12                    269.6n ± 4%   263.5n ± 0%  -2.24% (p=0.004 n=6)
PluginFilter/noop-wat/params:_real-12                     268.1n ± 1%   263.9n ± 0%  -1.53% (p=0.002 n=6)
PluginFilter/noop/params:_small-12                        323.4n ± 0%   319.6n ± 0%  -1.14% (p=0.002 n=6)
PluginFilter/noop/params:_real-12                         324.1n ± 0%   319.5n ± 0%  -1.42% (p=0.002 n=6)
PluginFilter/test/params:_small-12                        7.104µ ± 0%   7.173µ ± 1%  +0.98% (p=0.028 n=6)
PluginFilter/test/params:_real-12                         108.7µ ± 1%   109.0µ ± 0%       ~ (p=0.180 n=6)
PluginScore/noop-wat/params:_small-12                     267.4n ± 0%   265.7n ± 0%  -0.65% (p=0.002 n=6)
PluginScore/noop-wat/params:_real-12                      267.8n ± 0%   264.4n ± 1%  -1.25% (p=0.002 n=6)
PluginScore/noop/params:_small-12                         372.1n ± 0%   367.1n ± 1%  -1.34% (p=0.002 n=6)
PluginScore/noop/params:_real-12                          339.3n ± 0%   335.1n ± 0%  -1.25% (p=0.002 n=6)
PluginScore/test/params:_small-12                         4.083µ ± 0%   4.199µ ± 1%  +2.84% (p=0.002 n=6)
PluginScore/test/params:_real-12                          45.63µ ± 1%   46.60µ ± 1%  +2.12% (p=0.002 n=6)
PluginPrefilterFilterAndScore/noop-wat/params:_small-12   399.2n ± 0%   392.7n ± 1%  -1.63% (p=0.002 n=6)
PluginPrefilterFilterAndScore/noop-wat/params:_real-12    399.8n ± 0%   394.5n ± 0%  -1.31% (p=0.002 n=6)
PluginPrefilterFilterAndScore/noop/params:_small-12       555.2n ± 0%   556.5n ± 2%       ~ (p=0.937 n=6)
PluginPrefilterFilterAndScore/noop/params:_real-12        597.4n ± 1%   599.2n ± 1%       ~ (p=0.589 n=6)
PluginPrefilterFilterAndScore/test/params:_small-12       7.545µ ± 1%   7.657µ ± 0%  +1.49% (p=0.002 n=6)
PluginPrefilterFilterAndScore/test/params:_real-12        110.7µ ± 2%   113.0µ ± 1%  +2.03% (p=0.009 n=6)
geomean                                                   1.157µ        1.157µ       -0.03%
```