@codefromthecrypt 2023-06-26
ok using latest main, performance is 1-2.x% slower, but this update is about version coherence not performance.